transcribe-suite/requirements.txt
numpy>=1.26
soundfile>=0.12
pydub>=0.25
pyyaml>=6.0.1
tqdm>=4.66
rich>=13.7
huggingface-hub>=0.19
langdetect>=1.0.9

torch==2.8.0 ; platform_system == "Darwin" and platform_machine == "arm64"
torchaudio==2.8.0 ; platform_system == "Darwin" and platform_machine == "arm64"

whisperx==3.7.4
pyannote.audio>=3.3.2,<4.0.0
faster-whisper>=1.1.1
psutil>=5.9
